LC-sqrtx [Runtime: 1 ms; Memory: 41.1 MB]

+++ math/sqrtx.java
@@ -1,6 +1,5 @@
 class Solution {
     public int mySqrt(int x) {
-
         if (x == 0 || x == 1) {
             return x;
         }
@@ -8,9 +7,10 @@ class Solution {
         int left = 1;
         int right = x;
         int mid = -1;
+
         while (left <= right) {
             mid = left + (right - left) / 2;
-            if((long)mid * mid > (long)x) {
+            if ((long)mid * mid > x) {
                 right = mid - 1;
             } else if (mid * mid < x) {
                 left = mid + 1;
